Extracts rendering functionality out of Extension and into standalone class
